+++ core/src/test/java/org/jdbi/v3/core/mapper/ImmutablesTest.java
@@ -35,12 +35,14 @@ public class ImmutablesTest {
     @Rule
     public H2DatabaseRule dbRule = new H2DatabaseRule()
         .withConfig(JdbiImmutables.class, c -> c
-            .registerImmutable(SubValue.class)
-            .registerImmutable(FooBarBaz.class)
-            .registerModifiable(FooBarBaz.class)
-            .registerImmutable(Getter.class)
-            .registerImmutable(ByteArray.class)
-            .registerImmutable(DerivedProperty.class)
+            .registerImmutable(
+                    SubValue.class,
+                    FooBarBaz.class,
+                    Getter.class,
+                    ByteArray.class,
+                    DerivedProperty.class,
+                    Defaulty.class
+            ).registerModifiable(FooBarBaz.class)
         );
 
     private Jdbi jdbi;
@@ -232,5 +234,22 @@ public class ImmutablesTest {
 
     static class Boom extends RuntimeException {
         private static final long serialVersionUID = 1L;
+    }
+
+    @Test
+    public void testDefaultNotStoredInDb() {
+        assertThat(h.createQuery("select null as defaulted")
+                .mapTo(Defaulty.class)
+                .one())
+            .extracting(Defaulty::defaulted)
+            .isEqualTo(42);
+    }
+
+    @Value.Immutable
+    public interface Defaulty {
+        @Value.Default
+        default int defaulted() {
+            return 42;
+        }
     }
 }