+++ RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 - New Features
   - `ResultIterable<T>.map(Function<T, U>)` returns a `ResultIterable<U>` with elements transformed
     using the given mapper function.
+- Bug Fixes
+  - Immutables integration doesn't respect @Value.Default for primitives that are nulled in the db
 
 # 3.8.2
 - Improvements